Isolated report generation methods in their own file
