+++ source/app/business/reports/reporter.py
@@ -18,13 +18,10 @@
 #  along with this program; if not, write to the Free Software Foundation,
 #  Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 
-import base64
 import logging as log
 import os
 from datetime import datetime
 
-from app.business.errors import ObjectNotFoundError, BusinessProcessingError
-from app.datamgmt.case.case_db import get_case
 from app.iris_engine.access_control.iris_user import iris_current_user
 from docx_generator.docx_generator import DocxGenerator
 from docx_generator.exceptions import rendering_error
@@ -34,8 +31,6 @@ from app.business.cases import cases_export_to_json
 
 from app.datamgmt.activities.activities_db import get_auto_activities
 from app.datamgmt.activities.activities_db import get_manual_activities
-from app.iris_engine.module_handler.module_handler import call_modules_hook
-from app.iris_engine.utils.tracker import track_activity
 
 from app.models.models import CaseTemplateReport
 
@@ -214,63 +209,3 @@ class IrisMakeMdReport:
             return None, e.__str__()
 
         return output_file_path, 'Report generated'
-
-
-def generate_activities_report(caseid, report_id, safe_mode, tmp_dir):
-    call_modules_hook('on_preload_activities_report_create', data=report_id, caseid=caseid)
-    report = CaseTemplateReport.query.filter(CaseTemplateReport.id == report_id).first()
-    if not report:
-        raise ObjectNotFoundError()
-    # Get file extension
-    _, report_format = os.path.splitext(report.internal_reference)
-    # Depending on the template format, the generation process is different
-    if report_format == '.docx':
-        mreport = IrisMakeDocReport(tmp_dir, report_id, caseid, safe_mode)
-        fpath, logs = mreport.generate_doc_report('Activities')
-
-    elif report_format in ('.md', '.html'):
-        mreport = IrisMakeMdReport(tmp_dir, report_id, caseid, safe_mode)
-        fpath, logs = mreport.generate_md_report('Activities')
-
-    else:
-        raise BusinessProcessingError('Report error', data='Unknown report format.')
-    if fpath is None:
-        track_activity('failed to generate a report')
-        raise BusinessProcessingError('Failed to generate the report', data=logs)
-    call_modules_hook('on_postload_activities_report_create', data=report_id, caseid=caseid)
-    track_activity('generated a report')
-    return fpath
-
-
-def generate_investigation_report(caseid, report_id, safe_mode, tmp_dir):
-    call_modules_hook('on_preload_report_create', data=report_id, caseid=caseid)
-    report = CaseTemplateReport.query.filter(CaseTemplateReport.id == report_id).first()
-    if not report:
-        raise ObjectNotFoundError()
-    _, report_format = os.path.splitext(report.internal_reference)
-    if report_format in ('.md', '.html'):
-        mreport = IrisMakeMdReport(tmp_dir, report_id, caseid, safe_mode)
-        fpath, logs = mreport.generate_md_report('Investigation')
-
-    elif report_format == '.docx':
-        mreport = IrisMakeDocReport(tmp_dir, report_id, caseid, safe_mode)
-        fpath, logs = mreport.generate_doc_report('Investigation')
-
-    else:
-        raise BusinessProcessingError('Report error', data='Unknown report format.')
-    if fpath is None:
-        track_activity('failed to generate the report')
-        raise BusinessProcessingError('Failed to generate the report', data=logs)
-    with open(fpath, 'rb') as rfile:
-        encoded_file = base64.b64encode(rfile.read()).decode('utf-8')
-    res = get_case(caseid)
-    _data = {
-        'report_id': report_id,
-        'file_path': fpath,
-        'case_id': res.case_id,
-        'user_name': res.user.name,
-        'file': encoded_file
-    }
-    call_modules_hook('on_postload_report_create', data=_data, caseid=caseid)
-    track_activity('generated a report')
-    return fpath
